Add logic for forced password reset
Forcing a password reset on first login works like this:

There is a `reset_password` field in the users table. When the user
logs in, a middleware checks to see if that is true, and redirects the
user to change their password before they can continue to the rest of
the site.