Merge orders routes
There's a few loose ends here with this integration we'll need to fix
up. Namely, The patients, orders, and lab routes don't check for
any authorization other than basic log-in. We'll lock these down with
policies like we did for Medication and Users. Additionally, The split
administration routes have been merged, with a small controller used to
disambiguate them.