Move key decrypt to deployment script
PR tests don't have access to the environment variables needed to
decrypt the SSH key we use for deployment, so their builds will always
fail. Since we only need that key at deploy time, we can defer on
decrypting until that time. This commit also renames the deployment
script to make it more clear that it's intended for travis only.